[GPU] Refactored OV2.0 remote api for GPU
