+++ inference-engine/src/inference_engine/include/openvino/runtime/gpu/ocl/va.hpp
@@ -7,14 +7,14 @@
  * shared Video Acceleration device contexts
  * and shared memory tensors which contain Video Acceleration surfaces
  *
- * @file openvino/runtime/gpu/va.hpp
+ * @file openvino/runtime/gpu/ocl/va.hpp
  */
 #pragma once
 
 #include <memory>
 #include <string>
 
-#include "openvino/runtime/gpu/ocl.hpp"
+#include "openvino/runtime/gpu/ocl/ocl.hpp"
 
 // clang-format off
 #include <va/va.h>
@@ -23,6 +23,7 @@
 namespace ov {
 namespace runtime {
 namespace gpu {
+namespace ocl {
 
 /**
  * @brief This class represents an abstraction for GPU plugin remote tensor
@@ -92,10 +93,13 @@ public:
      * @brief Constructs remote context object from VA display handle
      * @param core OpenVINO Runtime Core object
      * @param device A `VADisplay` to create remote context from
+     * @param target_tile_id Desired tile id within given context for multi-tile system. Default value (-1) means
+     * that root device should be used
      */
-    VAContext(Core& core, VADisplay device) : ClContext(core, (cl_context) nullptr) {
+    VAContext(Core& core, VADisplay device, int target_tile_id = -1) : ClContext(core, (cl_context) nullptr) {
         ParamMap context_params = {{GPU_PARAM_KEY(CONTEXT_TYPE), GPU_PARAM_VALUE(VA_SHARED)},
-                                   {GPU_PARAM_KEY(VA_DEVICE), static_cast<gpu_handle_param>(device)}};
+                                   {GPU_PARAM_KEY(VA_DEVICE), static_cast<gpu_handle_param>(device)},
+                                   {GPU_PARAM_KEY(TILE_ID), target_tile_id}};
         *this = core.create_context(device_name, context_params);
     }
 
@@ -137,6 +141,7 @@ public:
         return create_tensor(type, shape, params);
     }
 };
+}  // namespace ocl
 }  // namespace gpu
 }  // namespace runtime
 }  // namespace ov
